Fixed Original Input display issues. Removed App.vue and HelloWorld.vue and replaced them with a single QuizGenerator.vue. Used Vetur to format the code
