+++ quiz-app-584/src/components/LoadingSpinner.vue
@@ -1,45 +1,49 @@
 <template>
-    <div class="loading-spinner">
-        <div class="spinner"></div>
-        <p>Loading...</p>
-    </div>
+  <div class="loading-spinner">
+    <div class="spinner"></div>
+    <p>Loading...</p>
+  </div>
 </template>
 
 <script>
-    export default {
-    name: 'LoadingSpinner',
-    };
+export default {
+  name: "LoadingSpinner",
+};
 </script>
   
 <style scoped>
 .loading-spinner {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    background: rgba(255, 255, 255, 0.8);
-    z-index: 1000; /* Ensure the spinner is on top of other elements */
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: rgba(255, 255, 255, 0.8);
+  z-index: 1000; /* Ensure the spinner is on top of other elements */
 }
 
 .spinner {
-    border: 8px solid #f3f3f3; /* Light grey */
-    border-top: 8px solid #3498db; /* Blue */
-    border-radius: 50%;
-    width: 50px;
-    height: 50px;
-    /* Use the defined rotation animation */
-    animation: spin 1s linear infinite;
+  border: 8px solid #f3f3f3; /* Light grey */
+  border-top: 8px solid #3498db; /* Blue */
+  border-radius: 50%;
+  width: 50px;
+  height: 50px;
+  /* Use the defined rotation animation */
+  animation: spin 1s linear infinite;
 }
 
 /* Define the rotation animation using keyframes  */
 @keyframes spin {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
 }
 </style>
